refactor(web): collapse QueuePanel's 3 duplicated row variants into QueueRow
