+++ apps/web/src/components/player/QueuePanel.tsx
@@ -1,13 +1,8 @@
-import { memo, useCallback, useRef, useState, useMemo } from 'react';
+import { useCallback, useRef, useState, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { usePlaybackStore } from '@/stores/usePlaybackStore';
-import type { Track } from '@/stores/types';
-import { cn } from '@/lib/utils';
-import { formatDuration } from '@shiranami/shared';
-import { X, Trash2, Music, Play, Pause, GripVertical } from 'lucide-react';
-import { EqBars } from '@/components/shared/EqBars';
-import { TrackThumbnail } from '@/components/shared/TrackThumbnail';
-import { motion } from 'motion/react';
+import { Trash2, Music } from 'lucide-react';
+import { SortableQueueRow, DragOverlayContent, QueueItem } from '@/components/player/QueueRow';
 import {
   DndContext,
   DragOverlay,
@@ -19,110 +14,7 @@ import {
   type DragStartEvent,
   type DragEndEvent,
 } from '@dnd-kit/core';
-import { SortableContext, verticalListSortingStrategy, useSortable } from '@dnd-kit/sortable';
-import { CSS } from '@dnd-kit/utilities';
-
-/* ── Sortable queue row (for Up Next items) ─────────────── */
-
-interface SortableQueueRowProps {
-  track: Track;
-  sortableId: string;
-  queueIndex: number;
-  onPlay: (queueIndex: number) => void;
-  onRemove: (e: React.MouseEvent, queueIndex: number) => void;
-}
-
-function SortableQueueRow({
-  track,
-  sortableId,
-  queueIndex,
-  onPlay,
-  onRemove,
-}: SortableQueueRowProps) {
-  const { t } = useTranslation('queue');
-  const { attributes, listeners, setNodeRef, transform, transition, isDragging } = useSortable({
-    id: sortableId,
-  });
-
-  const style = {
-    transform: CSS.Transform.toString(transform),
-    transition,
-  };
-
-  return (
-    <div
-      ref={setNodeRef}
-      style={style}
-      className={cn(
-        'flex items-center gap-1.5 px-2 py-1.5 rounded-lg transition-all duration-200 group cursor-pointer hover:bg-accent',
-        isDragging && 'opacity-30'
-      )}
-      onClick={() => onPlay(queueIndex)}
-    >
-      <button
-        className="shrink-0 p-0.5 rounded text-muted-foreground/20 opacity-0 group-hover:opacity-100 hover:text-muted-foreground/60 transition-all duration-150 cursor-grab active:cursor-grabbing touch-none"
-        aria-label={t('dragToReorder')}
-        {...attributes}
-        {...listeners}
-        onClick={e => e.stopPropagation()}
-      >
-        <GripVertical className="w-3 h-3" />
-      </button>
-
-      <TrackThumbnail
-        albumArt={track.albumArt}
-        alt={track.title}
-        imgClassName="rounded-md"
-        className="w-8 h-8 rounded-md bg-surface"
-        fallback={<Play className="w-3 h-3 text-muted-foreground/40" />}
-      />
-
-      <div className="min-w-0 flex-1">
-        <p className="text-xs font-medium truncate">{track.title}</p>
-        <p className="text-[10px] text-muted-foreground/50 truncate">{track.artist}</p>
-      </div>
-
-      <span className="text-[10px] text-muted-foreground/30 tabular-nums shrink-0">
-        {track.duration > 0 ? formatDuration(track.duration) : ''}
-      </span>
-
-      <motion.button
-        whileTap={{ scale: 0.75 }}
-        onClick={(e: React.MouseEvent) => onRemove(e, queueIndex)}
-        className="shrink-0 p-0.5 rounded text-muted-foreground/20 opacity-0 group-hover:opacity-100 hover:text-destructive transition-all duration-150"
-        aria-label={t('remove')}
-      >
-        <X className="w-3 h-3" />
-      </motion.button>
-    </div>
-  );
-}
-
-/* ── Drag overlay content ──────────────────────────────── */
-
-function DragOverlayContent({ track }: { track: Track }) {
-  return (
-    <div className="flex items-center gap-1.5 px-2 py-1.5 rounded-lg bg-accent text-foreground shadow-lg">
-      <div className="shrink-0 p-0.5 text-muted-foreground/40">
-        <GripVertical className="w-3 h-3" />
-      </div>
-      <TrackThumbnail
-        albumArt={track.albumArt}
-        alt={track.title}
-        imgClassName="rounded-md"
-        className="w-8 h-8 rounded-md bg-surface"
-        fallback={<Play className="w-3 h-3 text-muted-foreground/40" />}
-      />
-      <div className="min-w-0 flex-1">
-        <p className="text-xs font-medium truncate">{track.title}</p>
-        <p className="text-[10px] text-muted-foreground/50 truncate">{track.artist}</p>
-      </div>
-      <span className="text-[10px] text-muted-foreground/30 tabular-nums shrink-0">
-        {track.duration > 0 ? formatDuration(track.duration) : ''}
-      </span>
-    </div>
-  );
-}
+import { SortableContext, verticalListSortingStrategy } from '@dnd-kit/sortable';
 
 /* ── Main QueuePanel ───────────────────────────────────── */
 
@@ -289,75 +181,5 @@ export function QueuePanel() {
     </div>
   );
 }
-
-/* ── Now Playing item (not draggable) ──────────────────── */
-
-interface QueueItemProps {
-  track: { id: string; title: string; artist: string; albumArt?: string; duration: number };
-  index: number;
-  isActive: boolean;
-  isPlaying: boolean;
-  onPlay: (index: number) => void;
-  onRemove: (e: React.MouseEvent, index: number) => void;
-}
-
-const QueueItem = memo(function QueueItem({
-  track,
-  index,
-  isActive,
-  isPlaying,
-  onPlay,
-  onRemove,
-}: QueueItemProps) {
-  const { t } = useTranslation('queue');
-  return (
-    <div
-      className={cn(
-        'flex items-center gap-2.5 px-2 py-1.5 rounded-lg transition-all duration-200 group cursor-pointer',
-        isActive ? 'bg-primary/[0.08]' : 'hover:bg-accent'
-      )}
-      onClick={() => onPlay(index)}
-    >
-      <TrackThumbnail
-        albumArt={track.albumArt}
-        alt={track.title}
-        imgClassName="rounded-md"
-        className={cn('w-8 h-8 rounded-md', isActive ? 'bg-primary/15' : 'bg-surface')}
-        fallback={
-          isActive && isPlaying ? (
-            <>
-              <span className="sr-only">{t('nowPlaying')}</span>
-              <EqBars size="sm" />
-            </>
-          ) : isActive ? (
-            <Pause className="w-3 h-3 text-primary" />
-          ) : (
-            <Play className="w-3 h-3 text-muted-foreground/40" />
-          )
-        }
-      />
-
-      <div className="min-w-0 flex-1">
-        <p className={cn('text-xs font-medium truncate', isActive && 'text-primary')}>
-          {track.title}
-        </p>
-        <p className="text-[10px] text-muted-foreground/50 truncate">{track.artist}</p>
-      </div>
-
-      <span className="text-[10px] text-muted-foreground/30 tabular-nums shrink-0">
-        {track.duration > 0 ? formatDuration(track.duration) : ''}
-      </span>
-
-      <motion.button
-        whileTap={{ scale: 0.75 }}
-        onClick={(e: React.MouseEvent) => onRemove(e, index)}
-        className="shrink-0 p-0.5 rounded text-muted-foreground/20 opacity-0 group-hover:opacity-100 hover:text-destructive transition-all duration-150"
-        aria-label={t('remove')}
-      >
-        <X className="w-3 h-3" />
-      </motion.button>
-    </div>
-  );
-});
 
 export default QueuePanel;
